+++ Source/Krypton Components/Krypton.Toolkit/Controls Toolkit/KryptonDomainUpDown.cs
@@ -1617,6 +1617,26 @@ public class KryptonDomainUpDown : VisualControlBase,
     }
 
     /// <summary>
+    /// Raises the RightToLeftChanged event.
+    /// </summary>
+    /// <param name="e">An EventArgs containing event data.</param>
+    protected override void OnRightToLeftChanged(EventArgs e)
+    {
+        UpdateForRightToLeft();
+        base.OnRightToLeftChanged(e);
+    }
+
+    /// <summary>
+    /// Raises the <see cref="VisualControlBase.RightToLeftLayoutChanged"/> event.
+    /// </summary>
+    /// <param name="e">An EventArgs containing event data.</param>
+    protected override void OnRightToLeftLayoutChanged(EventArgs e)
+    {
+        UpdateForRightToLeft();
+        base.OnRightToLeftLayoutChanged(e);
+    }
+
+    /// <summary>
     /// Raises the EnabledChanged event.
     /// </summary>
     /// <param name="e">An EventArgs that contains the event data.</param>
@@ -2204,6 +2224,14 @@ public class KryptonDomainUpDown : VisualControlBase,
                 OnMouseLeave(e);
             }
         }
+    }
+
+    private void UpdateForRightToLeft()
+    {
+        _domainUpDown.RightToLeft = RightToLeft;
+        _domainUpDown.UpDownAlign = ToolkitRtlLayout.IsRtl(this)
+            ? LeftRightAlignment.Left
+            : LeftRightAlignment.Right;
     }
     #endregion
 }